third_party/blink/web_tests/external/wpt/css/css-ui/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property caret-shape has initial value auto
  assert_true: caret-shape doesn't seem to be supported in the computed style expected true got false
[FAIL] Property caret-shape inherits
  assert_true: caret-shape doesn't seem to be supported in the computed style expected true got false
[FAIL] Property nav-down has initial value auto
  assert_true: nav-down doesn't seem to be supported in the computed style expected true got false
[FAIL] Property nav-down does not inherit
  assert_true: expected true got false
[FAIL] Property nav-left has initial value auto
  assert_true: nav-left doesn't seem to be supported in the computed style expected true got false
[FAIL] Property nav-left does not inherit
  assert_true: expected true got false
[FAIL] Property nav-right has initial value auto
  assert_true: nav-right doesn't seem to be supported in the computed style expected true got false
[FAIL] Property nav-right does not inherit
  assert_true: expected true got false
[FAIL] Property nav-up has initial value auto
  assert_true: nav-up doesn't seem to be supported in the computed style expected true got false
[FAIL] Property nav-up does not inherit
  assert_true: expected true got false
[FAIL] Property user-select does not inherit
  assert_not_equals: got disallowed value "none"
Harness: the test ran to completion.
